requirements.txt
scipy
h5py
numpy
ipython
matplotlib
NearPy
Pillow
python-dateutil
PyYAML
scikit-image
scikit-learn
six
trimesh
pyhull
cvxopt
mayavi
pandas
visualization
pyquaternion
numba
autolab-core
autolab-perception
torch>=0.4.1
tensorboardX
